python_plugins/httpie_google_auth_plugin/requirements.txt
httpie
google-auth
